requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
sqlalchemy>=2.0
pydantic>=2.0
redis>=4.5
python-jose>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pytest>=7.0